ldobrovols ttys000  Nov 13 12:32

+++ src/mx_memset.c
@@ -3,7 +3,7 @@
 void *mx_memset(void *b, int c, size_t len) {
 	char *temp = (char*) b;
 
-    for (unsigned long i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
     	temp[i] = c;
 
     return (void*) temp;
